refactor: split core test classes by file
